requirements.txt
appnope==0.1.0
asttokens==2.2.1
backcall==0.1.0
bcrypt==4.0.1
blinker==1.6.2
click==8.1.6
DateTime==5.2
decorator==4.3.0
executing==1.2.0
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
ipython==8.14.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
ptyprocess==0.6.0
pure-eval==0.2.2
pycparser==2.19
Pygments==2.16.1
python-dateutil==2.7.3
pytz==2023.3
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==2.0.19
stack-data==0.6.2
text-unidecode==1.2
traitlets==5.9.0
typing_extensions==4.7.1
wcwidth==0.1.7
Werkzeug==2.3.6
WTForms==2.2.1
zope.interface==6.0
